celery_task/requirements.txt
amqp==5.2.0
asyncio==3.4.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
kombu==5.3.7
mysql-connector-python==9.0.0
prompt_toolkit==3.0.47
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
